// convex/_generated/server.d.ts
type Env = {
  readonly CONVEX_CLOUD_URL: string;
  readonly CONVEX_SITE_URL: string;
  readonly ADMIN_API_KEY: string | undefined;
  readonly RESEARCH_API_KEYS: string | undefined;
};
// ...
export declare const httpAction: HttpActionBuilder;

/**
 * Typesafe environment variables.
 *
 * This includes platform-provided env vars and any variables declared in
 * `convex.config.ts`.
 */
export declare const env: Env;